Change idle implementation to be graceful
Replace the use of net.Conn's SetDeadline() in the idling implementation
with a custom monitor that gracefully resets streams and closes the
connection. Abrupt termination of the connection using SetDeadline()
results in things such as proxy servers being slow to detect idled
connections, meaning clients connected to the proxy aren't immediately
notified of the connection closure.

Also incorporates some of the code from
8239520e2d16d25a398eab5eca5fe8157347c55f to fix the issue of being
unable to Reset() a stream if it has data waiting to be read.

Signed-off-by: Andy Goldstein <agoldste@redhat.com>