Fix possible data loss after goaway received
All frames are processed by workers running in goroutines. If 1 worker's
queue is backed up with data frames and another worker receives and
processes a goaway frame, it's a race as to whether or not the workers
will be able to process the pending data frames before the streams are
removed from the streams map.

This fix processes the goaway frame synchronously in Serve() instead of
in a worker goroutine. We now wait for all the workers to drain their
frame queues before tearing down the streams.