Make sure stream.Read() unblocks for close + reset
Make sure that stream.Read() always unblocks after a stream.Close() followed by a stream.Reset().
Even if the stream is considered finished (which happens when you call stream.Close()), we need to
still call stream.closeRemoteChannels() when stream.Reset() is called, regardless of the value of
stream.finished. This ensures that stream.closeChan is closed, which unblocks readers.